fix small math mistake and potential NameError in trajecotry interpolation code
dot_product represented the projection of the vector from the nearest neighbor to the IC onto the vector from the nearest neighbor to the second nearest neighbor. This is a distance! However, it was then tested if this quantity is <= 0.5, and an exception is raised if not. The idea is that it being too big means the other grid point was actually closer. However, that would only be true if we had already divided by the distance between the nearest neighbor and the second nearest neighbor (since that converts the projection into a fraction of the distance along the vector from nearest neighbor to second nearest neighbor). To resolve, I simply directly calculate that ratio (alpha) rather than the intermediate dot_product quantity.